PDFBOX-4437: refactor xml into main class
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1852275 13f79535-47bb-0310-9956-ffa450edef68
